apps/crawler/requirements.txt
requests==2.32.4
beautifulsoup4==4.13.4
sqlalchemy==2.0.41
python-dotenv==1.1.0
psycopg[binary]==3.2.9
apscheduler==3.10.4
playwright==1.53.0
